Make jsondb optional
